Newest Version Bugfix und Readme change

+++ UninstallAvira/UninstallAvira.ps1
@@ -15,15 +15,19 @@ Param(
     [switch]$Restart
 )
 
-Write-Host "Performing uninstallation of Avira Antivirus"
-Start-Process -FilePath "C:\Program Files\Avira\Antivirus\presetup.exe" -ArgumentList "/remsilent"
 
-$avira = Get-ItemProperty HKLM:\Software\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object {($_.Displayname -eq "Avira") -and ($_.QuietUninstallString -like '"C:\ProgramData\Package Cache\*\Avira.OE.Setup.Bundle.exe" /uninstall /quiet')}
-Write-Host "Performing uninstallation of Avira Launcher"
+if((Test-Path "C:\Program Files\Avira\Antivirus\presetup.exe") -eq $true){
+    Write-Host "Performing uninstallation of Avira Antivirus"
+    Start-Process -FilePath "C:\Program Files\Avira\Antivirus\presetup.exe" -ArgumentList "/remsilent"
+    Wait-Process -Name "presetup"
+    $avira = Get-ItemProperty HKLM:\Software\Wow6432Node\Microsoft\Windows\CurrentVersion\Uninstall\* | Where-Object {($_.Displayname -eq "Avira") -and ($_.QuietUninstallString -like '"C:\ProgramData\Package Cache\*\Avira.OE.Setup.Bundle.exe" /uninstall /quiet')}
+    Write-Host "Performing uninstallation of Avira Launcher"
+    Start-Process -FilePath $avira.BundleCachePath -Wait -ArgumentList "/uninstall /quiet"
+    
+    If ($Restart.IsPresent -eq $true){
+        Restart-Computer -Force
+    }
 
-Start-Process -FilePath $avira.BundleCachePath -Wait -ArgumentList "/uninstall /quiet"
-
-If ($Restart.IsPresent -eq $true){
-    Restart-Computer -Force
 }
+
